// front/src/api/generated/models/responseShortCategoryJson.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * ResidentialExpenses.API | v1
 * OpenAPI spec version: 1.0.0
 */
import type { CategoryPurposeJson } from './categoryPurposeJson';

export interface ResponseShortCategoryJson {
  id?: number;
  description?: string;
  purpose?: CategoryPurposeJson;
}