// front/src/api/generated/models/responseShortTransactionJson.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * ResidentialExpenses.API | v1
 * OpenAPI spec version: 1.0.0
 */
import type { TransactionTypeJson } from './transactionTypeJson';

export interface ResponseShortTransactionJson {
  id?: number;
  description?: string;
  value?: number;
  type?: TransactionTypeJson;
  categoryId?: number;
  personId?: number;
}